+++ NENEQUEST_MODOKI/src/ScoreCounter.h
@@ -11,7 +11,7 @@ public:
 	void Finalize() override {};
 	void Update() override;
 	void Draw() override;
-	int GetTotalScore() const;
+	unsigned long long GetTotalScore() const;
 	void AddScore(const int score);
 
 private:
@@ -19,7 +19,7 @@ private:
 	int mScoreFontHandle;	// スコアを表示するのに使用するフォント
 	int mStrFontHandle;	// "Score"という文字列を表示するのに使用するフォント
 	const int SCORE_STR_RX = 1150;	// スコアを表示する際の，スコアの文字列の右端のx座標，文字列の右寄せで使用する 	
-	int mTotalScore;	// 合計スコア
+	unsigned long long mTotalScore;	// 合計スコア
 	std::string mTotalScoreStr;	// 描画する用の文字列の合計スコア
 	int mTotalScoreStrW;	// 描画する際の文字列の合計スコアの横幅
 };
